tests(package): move and fix tests for running with Wallaby.js
